@@ -1,25 +1,28 @@
-import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
-import { AngularFireFunctions } from '@angular/fire/functions';
-import { MatDialog, MatPaginator, MatSort, MatTableDataSource } from '@angular/material';
-import { AdminService } from 'app/services/core/admin.service';
-import { PaymentsService } from 'app/services/payments.service';
-import { ReplaySubject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
-import { DaudiCustomer } from '../../../../models/Daudi/customer/Customer';
-import { EquityBulk } from '../../../../models/ipn/EquityBulk';
-import { CreatePaymentComponent } from '../../components/create-payment/create-payment.component';
-import { CustomerManagementComponent } from '../customer-management/customer-management.component';
+import { Component, OnDestroy, OnInit, ViewChild } from "@angular/core";
+import { AngularFireFunctions } from "@angular/fire/functions";
+import { DaudiPayment } from "app/models/payment/DaudiPayment";
+import { AdminService } from "app/services/core/admin.service";
+import { PaymentsService } from "app/services/payments.service";
+import { ReplaySubject } from "rxjs";
+import { takeUntil } from "rxjs/operators";
+import { DaudiCustomer } from "../../../../models/Daudi/customer/Customer";
+import { CreatePaymentComponent } from "../../components/create-payment/create-payment.component";
+import { CustomerManagementComponent } from "../customer-management/customer-management.component";
+import { MatPaginator } from '@angular/material/paginator';
+import { MatSort } from '@angular/material/sort';
+import { MatTableDataSource } from '@angular/material/table';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
-  selector: 'app-payments',
-  templateUrl: './payments.component.html',
-  styleUrls: ['./payments.component.scss']
+  selector: "app-payments",
+  templateUrl: "./payments.component.html",
+  styleUrls: ["./payments.component.scss"]
 })
 export class PaymentsComponent implements OnInit, OnDestroy {
-  displayedColumns: string[] = ['bank', 'mode', 'bankref', 'amount', 'name', 'phone'];
+  displayedColumns: string[] = ["bank", "mode", "bankref", "amount", "name", "phone"];
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
   @ViewChild(MatSort, { static: true }) sort: MatSort;
-  unprocesseddatasource = new MatTableDataSource<EquityBulk>();
+  unprocesseddatasource = new MatTableDataSource<DaudiPayment>();
   comopnentDestroyed: ReplaySubject<boolean> = new ReplaySubject<boolean>();
 
   constructor(
@@ -42,17 +45,17 @@ export class PaymentsComponent implements OnInit, OnDestroy {
     this.comopnentDestroyed.complete();
   }
 
-  attachpayment(payment: EquityBulk) {
+  attachpayment(payment: DaudiPayment) {
     const dialogRef = this.dialog.open(CustomerManagementComponent, {
-      width: '80%',
-      data: 'Attach'
+      width: "80%",
+      data: "Attach"
     });
     dialogRef.afterClosed().pipe(takeUntil(this.comopnentDestroyed)).subscribe((result: DaudiCustomer[] | null) => {
       if (result) {
-        payment.billNumber = result[0].Id;
+        payment.transaction.billNumber = result[0].Id;
         payment.daudiFields.status = 48;
         payment.daudiFields.approvedby = this.adminservice.createUserObject();
-        this.functions.httpsCallable('ipnCallable')(payment)
+        this.functions.httpsCallable("ipnCallable")(payment)
           .pipe(takeUntil(this.comopnentDestroyed))
           .subscribe(res => {
             console.log(res);
@@ -65,8 +68,8 @@ export class PaymentsComponent implements OnInit, OnDestroy {
 
   addPayment() {
     this.dialog.open(CreatePaymentComponent, {
-      width: '80%',
-      data: 'Attach'
+      width: "80%",
+      data: "Attach"
     });
   }
 }
