@@ -1,44 +1,48 @@
-import {animate, sequence, state, style, transition, trigger} from '@angular/animations';
-import {Component, OnInit, ViewChild} from '@angular/core';
-import {MatDialog, MatPaginator, MatSort, MatTableDataSource} from '@angular/material';
-import {ActivatedRoute} from '@angular/router';
-import {ColumnsCustomizerComponent} from 'app/components/columns-customizer/columns-customizer.component';
-import {SendMsgComponent} from 'app/components/send-msg/send-msg.component';
-import {CoreService} from 'app/services/core/core.service';
-import {DepotService} from 'app/services/core/depot.service';
-import {ExcelService} from 'app/services/excel-service.service';
+import { animate, sequence, state, style, transition, trigger } from '@angular/animations';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatSort } from '@angular/material/sort';
+import { MatTableDataSource } from '@angular/material/table';
+import { ActivatedRoute } from '@angular/router';
+import { ColumnsCustomizerComponent } from 'app/components/columns-customizer/columns-customizer.component';
+import { SendMsgComponent } from 'app/components/send-msg/send-msg.component';
+import { CoreService } from 'app/services/core/core.service';
+import { DepotService } from 'app/services/core/depot.service';
+import { ExcelService } from 'app/services/excel-service.service';
 import * as moment from 'moment';
-import {ReplaySubject} from 'rxjs';
-import {takeUntil} from 'rxjs/operators';
-import {Order} from '../../models/Daudi/order/Order';
-import {emptytruck, Truck} from '../../models/Daudi/order/truck/Truck';
-import {SMS} from '../../models/Daudi/sms/sms';
-import {MyTimestamp} from '../../models/firestore/firestoreTypes';
+import { ReplaySubject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
+import { Order } from '../../models/Daudi/order/Order';
+import { emptytruck, Truck } from '../../models/Daudi/order/truck/Truck';
+import { SMS } from '../../models/Daudi/sms/sms';
+import { MyTimestamp } from '../../models/firestore/firestoreTypes';
+import { TooltipPosition } from '@angular/material/tooltip';
 
 @Component({
-  selector: 'trucks-table',
-  templateUrl: './trucks-table.component.html',
-  styleUrls: ['./trucks-table.component.scss'],
+  selector: "trucks-table",
+  templateUrl: "./trucks-table.component.html",
+  styleUrls: ["./trucks-table.component.scss"],
   animations: [
-    trigger('flyIn', [
-      state('in', style({transform: 'translateX(0)'})),
-      transition('void => *', [
-        style({height: '*', opacity: '0', transform: 'translateX(-550px)', 'box-shadow': 'none'}),
+    trigger("flyIn", [
+      state("in", style({ transform: "translateX(0)" })),
+      transition("void => *", [
+        style({ height: "*", opacity: "0", transform: "translateX(-550px)", "box-shadow": "none" }),
         sequence([
-          animate('.20s ease', style({height: '*', opacity: '.2', transform: 'translateX(0)', 'box-shadow': 'none'})),
-          animate('.15s ease', style({height: '*', opacity: 1, transform: 'translateX(0)'}))
+          animate(".20s ease", style({ height: "*", opacity: ".2", transform: "translateX(0)", "box-shadow": "none" })),
+          animate(".15s ease", style({ height: "*", opacity: 1, transform: "translateX(0)" }))
         ])
       ])
     ]),
-    trigger('detailExpand', [
-      state('collapsed', style({height: '0px', minHeight: '0', display: 'none'})),
-      state('expanded', style({height: '*'})),
-      transition('expanded <=> collapsed', animate('225ms cubic-bezier(0.4, 0.0, 0.2, 1)'))
+    trigger("detailExpand", [
+      state("collapsed", style({ height: "0px", minHeight: "0", display: "none" })),
+      state("expanded", style({ height: "*" })),
+      transition("expanded <=> collapsed", animate("225ms cubic-bezier(0.4, 0.0, 0.2, 1)"))
     ])]
 })
 export class TrucksTableComponent implements OnInit {
-  position = 'above'; // for tooltip
-
+  position: TooltipPosition = "above"; // for tooltip
+  expandedElement;
   orders: any[];
   dialogProperties: object = {};
   dialogsections: number[] = [];
@@ -47,14 +51,15 @@ export class TrucksTableComponent implements OnInit {
 
   trucksdataSource = new MatTableDataSource<Truck>();
 
-  truckcolumns = ['truckId', 'time', 'Print Status', 'Phone', 'orderCompanyName', 'driverName', 'driverId', 'truckReg', 'pmsQty', 'agoQty', 'ikQty'];
+  truckcolumns = ["truckId", "orderCompanyName", "time", "Print Status", "Phone", "driverName", "driverId", "truckReg", "pmsQty", "agoQty", "ikQty"];
 
   temporder = {};
   loadingtrucks = true;
   comopnentDestroyed: ReplaySubject<boolean> = new ReplaySubject<boolean>();
+  typedValue: string;
 
-  @ViewChild(MatPaginator, {static: true}) paginator: MatPaginator;
-  @ViewChild(MatSort, {static: true}) sort: MatSort;
+  @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
+  @ViewChild(MatSort, { static: true }) sort: MatSort;
 
   constructor(
     private dialog: MatDialog,
@@ -88,12 +93,12 @@ export class TrucksTableComponent implements OnInit {
     const dialogRef = this.dialog.open(ColumnsCustomizerComponent,
       {
         data: emptytruck,
-        width: '70%',
-        role: 'dialog'
+        width: "70%",
+        role: "dialog"
       });
     dialogRef.afterClosed().pipe(takeUntil(this.comopnentDestroyed)).subscribe(result => {
       if (result) {
-        this.excelService.exportAsExcelFile(this.trucksdataSource.data, 'Orders');
+        this.excelService.exportAsExcelFile(this.trucksdataSource.data, "Orders");
       }
     });
   }
@@ -110,10 +115,10 @@ export class TrucksTableComponent implements OnInit {
       company: clickedOrder.customer,
       type: {
         reason: null,
-        origin: 'custom'
+        origin: "custom"
       },
       contact: clickedOrder.customer.contact,
-      greeting: 'Jambo',
+      greeting: "Jambo",
       msg: null,
       status: {
         delivered: false,
@@ -122,9 +127,9 @@ export class TrucksTableComponent implements OnInit {
       timestamp: MyTimestamp.now()
     };
     const dialogRef = this.dialog.open(SendMsgComponent, {
-      role: 'dialog',
+      role: "dialog",
       data: sms,
-      height: 'auto'
+      height: "auto"
     });
     // this.dialog.open(SendMsgComponent);
   }
