@@ -1,15 +1,15 @@
-import {Injectable} from '@angular/core';
-import {AngularFirestore} from '@angular/fire/firestore';
-import {AngularFireMessaging} from '@angular/fire/messaging';
-import {Admin} from 'app/models/Daudi/admin/Admin';
-import {FCM} from 'app/models/Daudi/notification/FCM';
-import {NotificationService} from 'app/shared/services/notification.service';
-import {BehaviorSubject} from 'rxjs';
-import {distinctUntilChanged} from 'rxjs/operators';
-import {AdminService} from './core/admin.service';
+import { Injectable } from "@angular/core";
+import { AngularFirestore } from "@angular/fire/firestore";
+import { AngularFireMessaging } from "@angular/fire/messaging";
+import { Admin } from "app/models/Daudi/admin/Admin";
+import { FCM } from "app/models/Daudi/notification/FCM";
+import { NotificationService } from "app/shared/services/notification.service";
+import { BehaviorSubject } from "rxjs";
+import { distinctUntilChanged } from "rxjs/operators";
+import { AdminService } from "./core/admin.service";
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: "root"
 })
 export class FcmService {
   currentMessage = new BehaviorSubject(null);
@@ -31,7 +31,7 @@ export class FcmService {
 
   updateusertokes(user: Admin, token) {
     if (user.config.fcm.tokens.web !== token) {
-      this.db.firestore.collection('admins').doc(user.Id).update({'fcmtokens.web': token});
+      this.db.firestore.collection("admins").doc(user.Id).update({ "fcmtokens.web": token });
     }
   }
 
@@ -41,7 +41,7 @@ export class FcmService {
         this.updateusertokes(user, token);
       },
       (err) => {
-        console.error('Unable to get permission to notify.', err);
+        console.error("Unable to get permission to notify.", err);
       }
     );
   }
@@ -51,16 +51,16 @@ export class FcmService {
       (payload: FCM) => {
         // console.log('new message received. ', payload);
         switch (payload.notification.title) {
-          case 'Payment Received':
+          case "Payment Received":
             return this.notification.notify({
-              alert_type: 'cash',
+              alert_type: "cash",
               duration: 5000,
               title: payload.notification.title,
               body: payload.notification.body
             });
-          case 'Unprocessed Payment':
+          case "Unprocessed Payment":
             return this.notification.notify({
-              alert_type: 'unprocessedpayment',
+              alert_type: "unprocessedpayment",
               duration: 5000,
               title: payload.notification.title,
               body: payload.notification.body
