@@ -7,14 +7,16 @@ export interface StockQty {
   total: number;
   /**
    * The total that has been loaded, excluding amounts accumulated
+   * This is the single source of truth then calculating the qty available, and it must be kept up to date accoringly
    */
   used: number;
   /**
    * Quantity transferred to private depots
+   * Also update the qty used dusring a transfer
    */
   transferred: {
     total: number;
-    transfers: { depotId: string; entryId: string, qty: number }[];
+    transfers: { depotId: string; entryId: string, qty: number; }[];
   };
   /**
    * Qty loaded(drawn) directly at a depot
@@ -28,7 +30,7 @@ export interface StockQty {
      * quantities accumulate over time as a result of Observed vs Actual Quantities when loading at the depot
      */
     accumulated: number;
-  }
+  };
 }
 
 export const EmptyStockQty: StockQty = {
