@@ -1,12 +1,12 @@
-import {deepCopy} from '../../utils/deepCopy';
-import {emptymetadata, Metadata} from '../universal/Metadata';
+import { deepCopy } from '../../utils/deepCopy';
+import { AssociatedUser, EmptyAssociatedUser } from '../admin/AssociatedUser';
 
 export interface TaxExempt {
   amount: number;
-  metadata: Metadata;
+  user: AssociatedUser;
 }
 
 export const emptytaxExempt: TaxExempt = {
   amount: 0,
-  metadata: deepCopy<Metadata>(emptymetadata)
+  user: deepCopy<AssociatedUser>(EmptyAssociatedUser),
 };
