@@ -1,18 +1,7 @@
-import {deepCopy} from '../../utils/deepCopy';
-import {AssociatedUser, EmptyAssociatedUser} from '../admin/AssociatedUser';
-import {FuelType} from '../fuel/FuelType';
+import { Price, emptyprice } from '../depot/Price';
 
-export interface AvgPrice {
-  Id: string;
-  price: number;
-  user: AssociatedUser;
-  fueltytype: FuelType;
-  omcId?: string;
+export interface AvgPrice extends Price {
+  omcId: string;
 }
 
-export const emptyprice: AvgPrice = Object.freeze({
-  price: null,
-  Id: null,
-  user: deepCopy<AssociatedUser>(EmptyAssociatedUser),
-  fueltytype: null
-});
+export const emptyAvgprice: AvgPrice = { ...emptyprice, ...{ omcId: null } }
