@@ -1,6 +1,6 @@
-import {LocationStrategy, PathLocationStrategy} from '@angular/common';
-import {ErrorHandler, Injectable, Injector} from '@angular/core';
-import * as StackTrace from 'stacktrace-js';
+import { LocationStrategy, PathLocationStrategy } from '@angular/common';
+import { ErrorHandler, Injectable, Injector } from '@angular/core';
+import { fromError } from 'stacktrace-js';
 
 // import { LoggingService } from './services/logging-service.service';
 
@@ -17,14 +17,14 @@ export class MyErrorHandler implements ErrorHandler {
     const url = location instanceof PathLocationStrategy
       ? location.path() : '';
     // get the stack trace, lets grab the last 10 stacks only
-    StackTrace.fromError(error).then(stackframes => {
+    fromError(error).then(stackframes => {
       const stackString = stackframes
         .splice(0, 20)
-        .map(function(sf) {
+        .map(function (sf) {
           return sf.toString();
         }).join('\n');
       // log on the server
-      loggingService.log({message, url, stack: stackString});
+      loggingService.log({ message, url, stack: stackString });
     });
 
   }
