@@ -6,11 +6,10 @@ import {
   OnInit,
   Output
 } from "@angular/core";
-import { Validators } from "@angular/forms";
+import { Validators, FormGroup, FormControl, FormBuilder } from "@angular/forms";
 import { CoreService } from "app/services/core/core.service";
 import { CustomerService } from "app/services/customers.service";
 import { isEqual } from "lodash";
-import { FormBuilder, FormControl, FormGroup } from "ngx-strongly-typed-forms";
 import { ReplaySubject, Subject } from "rxjs";
 import { delay, takeUntil } from "rxjs/operators";
 import { CustomerDetail } from "../../../../models/Daudi/customer/CustomerDetail";
@@ -28,7 +27,7 @@ export class ContactFormComponent implements OnInit, OnChanges {
 
   @Input() newOrder: boolean;
 
-  @Output() modifiedOrder = new EventEmitter<Order>();
+  @Output() initDataChange = new EventEmitter<Order>();
   @Output() formValid = new EventEmitter<boolean>();
   /**
    * Emits whenever a compay is selected from the dropdown
@@ -48,15 +47,15 @@ export class ContactFormComponent implements OnInit, OnChanges {
   comopnentDestroyed: ReplaySubject<boolean> = new ReplaySubject<boolean>();
   subscriptions: Map<string, any> = new Map<string, any>();
 
-  contactForm: FormGroup<OrderContactForm> = new FormGroup<OrderContactForm>({
-    kraPin: new FormControl<string>("", [Validators.required]),
-    email: new FormControl<string>("", [Validators.required, Validators.email]),
-    companyName: new FormControl<string>("", [Validators.required]),
-    name: new FormControl<string>("", [
+  contactForm: FormGroup = new FormGroup({
+    kraPin: new FormControl("", [Validators.required]),
+    email: new FormControl("", [Validators.required, Validators.email]),
+    companyName: new FormControl("", [Validators.required]),
+    name: new FormControl("", [
       Validators.required,
       Validators.minLength(4)
     ]),
-    phone: new FormControl<string>("", [
+    phone: new FormControl("", [
       Validators.required,
       Validators.pattern("[0-9].{8}")
     ])
@@ -130,6 +129,8 @@ export class ContactFormComponent implements OnInit, OnChanges {
           this.formValid.emit(true);
         }
       });
+    this.initDataChange.emit(this.initData);
+
     this.core.loaders.customers
       .pipe(takeUntil(this.comopnentDestroyed))
       .subscribe(value => {
@@ -141,7 +142,7 @@ export class ContactFormComponent implements OnInit, OnChanges {
     /**
      * Silently update the related values
      */
-    console.log(selectedcompany)
+    console.log(selectedcompany);
     this.contactForm.controls.kraPin.setValue(selectedcompany.krapin, {
       emitEvent: false
     });
@@ -169,14 +170,14 @@ export class ContactFormComponent implements OnInit, OnChanges {
     this.selectedCustomer_ = detail;
     // const kraModified = this.initData ? this.initData.customer.krapin === values.kraPin : false;
     this.initData.customer = detail;
-    // this.formValid.emit(this.contactForm.valid);
-    // this.modifiedOrder.emit(this.initData);
+    this.formValid.emit(this.contactForm.valid);
+    this.initDataChange.emit(this.initData);
   }
 
   ngOnChanges(changes: any) {
     if (!this.newOrder && this.initData.customer.Id) {
       this.contactForm.disable();
-      this.companyselect(this.initData.customer)
+      this.companyselect(this.initData.customer);
     }
   }
 
