@@ -1,6 +1,6 @@
-import {Injectable} from '@angular/core';
-import {AngularFirestore} from '@angular/fire/firestore';
-import {FuelType} from 'app/models/Daudi/fuel/FuelType';
+import { Injectable } from '@angular/core';
+import { AngularFirestore } from '@angular/fire/firestore';
+import { FuelType } from 'app/models/Daudi/fuel/FuelType';
 import * as moment from 'moment';
 
 @Injectable({
@@ -12,15 +12,10 @@ export class PricesService {
 
   }
 
-  createavgprice(omcId: string) {
-    return this.avgPricesCollection(omcId)
-      .doc();
-  }
-
   avgPricesCollection(omcId: string) {
     return this.db.firestore.collection('omc')
       .doc(omcId)
-      .collection(`avgprice`);
+      .collection(`avgPrices`);
   }
 
   deleteavgprice(omcId: string, avgId: string) {
@@ -36,7 +31,17 @@ export class PricesService {
   priceCollection(omcId: string) {
     return this.db.firestore.collection('omc')
       .doc(omcId)
-      .collection(`price`);
+      .collection(`prices`);
+  }
+  minPriceCollection(omcId: string) {
+    return this.db.firestore.collection('omc')
+      .doc(omcId)
+      .collection(`minPrices`);
+  }
+  TaxCollection(omcId: string) {
+    return this.db.firestore.collection('omc')
+      .doc(omcId)
+      .collection(`taxExcemptPrices`);
   }
 
   getAvgpricesrange(omcId: string, depotId: string, fueltye: FuelType, start, stop?) {
@@ -50,8 +55,4 @@ export class PricesService {
     return query;
   }
 
-  createprice(omcId: string) {
-    return this.priceCollection(omcId)
-      .doc(this.db.createId());
-  }
 }
